+++ javaScript/projects/binary-searh-trees/src/containers/mergesort.js
@@ -1,0 +1,44 @@
+const mergeSort = (arr) => {
+  // console.log(arr);
+  if (arr.length === 1) return arr;
+  const midPoint = Math.floor(arr.length / 2);
+
+  const left = arr.slice(0, midPoint);
+  const right = arr.slice(midPoint);
+  const sortedLeft = mergeSort(left);
+  const sortedRight = mergeSort(right);
+  const sortedArr = [];
+
+  let i = 0;
+  let j = 0;
+  let k = 0;
+
+  while (i < sortedLeft.length && j < sortedRight.length) {
+    if (sortedLeft[i] < sortedRight[j]) {
+      sortedArr[k] = sortedLeft[i];
+      i += 1;
+    } else if (sortedLeft[i] > sortedRight[j]) {
+      sortedArr[k] = sortedRight[j];
+      j += 1;
+    } else {
+      sortedArr[k] = sortedRight[j];
+      i += 1;
+      j += 1;
+    }
+    k += 1;
+  }
+
+  for (; i < sortedLeft.length; i += 1) {
+    sortedArr[k] = sortedLeft[i];
+    k += 1;
+  }
+
+  for (; j < sortedRight.length; j += 1) {
+    sortedArr[k] = sortedRight[j];
+    k += 1;
+  }
+
+  return sortedArr;
+};
+
+export default mergeSort;