+++ javaScript/projects/binary-searh-trees/src/containers/node.js
@@ -2,6 +2,6 @@ export default class Node {
   constructor(data) {
     this.leftNode = null;
     this.rightNode = null;
-    this.data = !data ? null : data;
+    this.data = data !== null ? data : null;
   }
 }