+++ javaScript/projects/binary-searh-trees/src/containers/tree.js
@@ -1,8 +1,10 @@
+import mergeSort from './mergesort';
 import Node from './node';
 
 const Tree = (array) => {
   // Accepts an array when initialized.
   // The Tree class should have a root attribute, which uses the return value of buildTree
+  const sortedArray = mergeSort(array);
   let root;
   let memo = {};
   const buildTree = (arr, start, end) => {
@@ -12,63 +14,126 @@ const Tree = (array) => {
     // The buildTree function should return the level-0 root node.
     if (start > end) return null;
 
-    // const midPoint = Math.floor((start + end) / 2);
-    // const rootNode = new Node(arr[midPoint]);
-    // rootNode.leftNode = buildTree(arr, start, midPoint - 1);
-    // rootNode.rightNode = buildTree(arr, midPoint + 1, end);
-    const rootNode = new Node(arr[0]);
-    for (let i = 1; i < arr.length; i += 1) {
-      let node = rootNode;
-      while (node) {
-        if (node.data < arr[i]) {
-          if (node.rightNode === null) {
-            node.rightNode = new Node(arr[i]);
-            node = null;
-          } else {
-            node = node.rightNode;
-          }
-        } else if (node.data > arr[i]) {
-          if (node.leftNode === null) {
-            node.leftNode = new Node(arr[i]);
-            node = null;
-          } else {
-            node = node.leftNode;
-          }
-        } else {
-          node = null;
-        }
-      }
-    }
+    const midPoint = Math.floor((start + end) / 2);
+    const rootNode = new Node(arr[midPoint]);
+    rootNode.leftNode = buildTree(arr, start, midPoint - 1);
+    rootNode.rightNode = buildTree(arr, midPoint + 1, end);
+
     return rootNode;
   };
 
   const insertNode = (value) => {
+    let node = root;
+    const newNode = new Node(value);
+    while (node) {
+      if (node.data < value) {
+        // go right
+        if (node.rightNode === null) {
+          node.rightNode = newNode;
+        } else {
+          node = node.rightNode;
+        }
+      } else if (node.data > value) {
+        // go left
+        if (node.leftNode === null) {
+          node.leftNode = newNode;
+        } else {
+          node = node.leftNode;
+        }
+      } else {
+        // duplicate found
+        break;
+      }
+    }
     // Implementation of these methods should traverse the tree and manipulate the nodes and their connections.
+  };
+
+  const find = (value) => {
+    // Accepts a value and returns the node with the given value.
+    let node = root;
+    while (node) {
+      if (node.data > value) {
+        // go left
+        node = node.leftNode;
+      } else if (node.data < value) {
+        // go right
+        node = node.rightNode;
+      } else {
+        break;
+      }
+    }
+    return node;
+  };
+
+  const predecessor = (value) => {
+    // returns predecessor node of given value
+    // how to refactor this?
+    let node = root;
+
+    while (node) {
+      if (node.data > value) {
+        // go left
+        node = node.leftNode;
+      } else if (node.data < value) {
+        // go right
+        node = node.rightNode;
+      } else {
+        break;
+      }
+
+      if (
+        (node.leftNode && node.leftNode.data === value) ||
+        (node.rightNode && node.rightNode.data === value)
+      ) {
+        break;
+      }
+    }
+    return node;
   };
 
   const deleteNode = (value) => {
     // Implementation of these methods should traverse the tree and manipulate the nodes and their connections.
     // There will be several cases for delete, such as when a node has children or not.
+    const targetNode = find(value);
+    const parentNode = predecessor(value);
+    const { leftNode } = targetNode;
+    const { rightNode } = targetNode;
+    console.log(targetNode);
+    if (leftNode && rightNode) {
+      // has 2 children
+      console.log(`targetNode has 2 children`);
+    } else if (leftNode || rightNode) {
+      // has 1 child
+      console.log(`targetNode has 1 child`);
+    } else {
+      // no children
+      console.log(`targetNode has no children`);
+    }
   };
 
-  const find = (value) => {
-    // Accepts a value and returns the node with the given value.
-  };
-
-  const levelOrder = (callback) => {
+  const levelOrder = (callback, node = root, arr = []) => {
     // Accepts a random OPTIONAL callback function as its parameter
     // Traverse the tree in breadth-first level order and provide each node as an argument to the callback.
     // The callback will perform an operation on each node following the order in which they are traversed.
     // The method should return an array of values if no callback is given as an argument.
     // You will want to use an array acting as a queue to keep track of all the child nodes that you have yet to traverse and to add new ones to the list
+    return `levelOrder()`;
   };
 
-  const inOrder = (callback) => {
+  const inOrder = (callback, node = root, arr = []) => {
     // left => root => right
     // elements of the array will be in order
     // Accepts a random optional callback as a parameter.
     // Traverse the tree in their respective depth-first order and yield each node to the provided callback.
     // Return an array of values if no callback is given as an argument.
+
+    if (node) {
+      inOrder(callback, node.leftNode, arr);
+      arr.push(node.data);
+      inOrder(callback, node.rightNode, arr);
+    }
+
+    return arr;
   };
 
   const preOrder = (callback) => {
@@ -118,11 +183,14 @@ const Tree = (array) => {
     }
   };
 
-  root = buildTree(array, 0, array.length - 1);
-  buildTree(array, 0, array.length - 1);
+  root = buildTree(sortedArray, 0, sortedArray.length - 1);
   return {
-    buildTree,
     root,
+    buildTree,
+    insertNode,
+    deleteNode,
+    find,
+    inOrder,
     prettyPrint,
   };
 };